Update numpy version requirement
The LCO Python packages upon which this code depends have packaging
bugs, which mean that they cannot be installed without the user having
to explicitly install numpy first.

The numpy version installed by the Dockerfile and the numpy version
installed as part of the dependencies specified in requirements.txt
should be the same version to save significant amounts of time when
building the Docker image.

In the longer term, the LCO Python packages should be fixed.